Fix: "Order and Book from Double to BigDecimal"

+++ src/main/java/com/andersentask/bookshop/console/Commands.java
@@ -10,8 +10,13 @@ import com.andersentask.bookshop.order.enums.OrderStatus;
 import com.andersentask.bookshop.request.entities.Request;
 import com.andersentask.bookshop.utils.serialization.RepositorySerializer;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
-import java.util.*;
+import java.util.Comparator;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
 
 
 public class Commands {
@@ -184,7 +189,7 @@ public class Commands {
      * @param endOfPeriod   end of the period
      * @return double income for the chosen period
      */
-    public double getIncomeForPeriod(LocalDateTime startOfPeriod, LocalDateTime endOfPeriod) {
+    public BigDecimal getIncomeForPeriod(LocalDateTime startOfPeriod, LocalDateTime endOfPeriod) {
         return appContextConfig.getOrderService().getIncomeForPeriod(startOfPeriod, endOfPeriod);
     }
